Most recent block is opening window after auto-login
Closes #120
