--- frontend/src/components/Block.jsx
@@ -1,13 +1,0 @@
-import React, { useEffect } from 'react';
-import Day from './Day';
-
-export default function Block({ currentWeekIndex, currentDayIndex, blockData, addExerciseToDay, addSetToExercise, deleteSetFromExercise, updateSetData, moveExercise }) {
-
-
-
-  return (
-    <>
-      
-    </>
-  );
-}
